update SILO_grid_sum
Testing again with a 3rd dataset 1985-1989 (the maximum number of rows supported by excel) and it seems to works without any hiccups.

Can't seem to pin point the source of the issue with the larger csv, might be easier at this stage to just abandon the "SILO-1985-2020.csv" and update this script to calculate the cum_sum for each year individually and only compile all the years together at the very end.

If the glitch is due to the large number of rows, I'm hoping that processing each year separately will avoid it - the script should also flag if the error is happening in the future.