Update SILO_grid_sum_V2.py
Ran another comparison between excel and to_csv() for exactly the same number of rows, seems the error must be something with the encoding or rounding that to_csv() isn't doing that excel is. Next test will be trying to do the same rounding that excel does but from python.