SILO_grid_sum_v3 fix
overhauled the v2 script, made a new version so that I could compare.

This one replaces most of the use of csvReader with just pd.read_csv() and it much faster. Also allowed me to much more easily fix the issue with date formatting that I was having when using the data in the calibration script as well as spotting an error with that previous data where it was just printing the full list of dates from the unprocessed dataset (i.e. n grids meant n repeats of the same date ).

Time to processes MUCH faster, output files also much smaller and in the correct format.